Exception handler delete

+++ src/main/java/com/lucasstore/store/services/UserService.java
@@ -2,8 +2,11 @@ package com.lucasstore.store.services;
 
 import com.lucasstore.store.models.User;
 import com.lucasstore.store.repositories.UserRepository;
+import com.lucasstore.store.services.exceptions.DatabaseException;
 import com.lucasstore.store.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Integer id){
-        userRepository.deleteById(id);
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+
+        try{
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Integer id, User user){
